requirements.txt
Eve
eve-healthcheck
Eve-Swagger
Eve-SQLAlchemy
eve-auth-jwt-fork
HubSpot3
celery[sqs]
celery
boto3
cloudwatch
